@@ -2,7 +2,7 @@ package com.github.lulewiczg.controller.ui.component;
 
 import android.app.Activity;
 import android.content.Context;
-import android.support.v4.view.ViewPager;
+import androidx.viewpager.widget.ViewPager;
 import android.util.AttributeSet;
 import android.view.MotionEvent;
 
