@@ -1,6 +1,6 @@
 package com.github.lulewiczg.controller.ui.fragment;
 
-import android.support.v4.app.Fragment;
+import androidx.fragment.app.Fragment;
 
 /**
  * Created by Grzegurz on 2016-03-07.
